Completed authentication with cookies.
the problem i had(at least from what i understood) was that i was setting the cookie in the backend, something i shouldve done in the front end. I just assumed that wherever the cookie would be, it would work but it didnt. So i fixed that and set the cookie in the front end.